MOBILE WALLET AND SYNCE ONLINE WALLET
Hi - I just wanted to let you know that I couldn't withdraw my VET from the mobile wallet that I downloaded on my Iphone so I uninstalled and re installed a new wallet using my seed words and my original password and there is nothing there.  When I tried to withdraw my VET from the mobile wallet, I put in the address to the Sync wallet - the mobile wallet didn't say that it sent the tokens because it kept saying my password was incorrect.

So, now, there is nothing in the mobile wallet and nothing in the Sync wallet.  What happened to my VET tokens?  Could someone retrieve them for me?  This wasn't my fault.